requirements.txt
streamlit
sentence-transformers
faiss-cpu
pandas
numpy
pypdf
python-dotenv
openai
pytest
